v1.1.0
This release marks the beginning of the history of this repo.
Prior to this, the repo was littered with test commits made when
developing the CI pipeline. This really should have been in a different
repo but, oh well, here we are. The PyPI and NPM packages have been
published with prior versions and there's no undoing that. So here's
to the future. 🌱