feat!: expose API runtime constructs (#167)
Introduce separate constructs for just service lambda runtimes, without the API Gateway.

BREAKING CHANGE: This update will change the internal node IDs of the constructs. This will cause the Lambdas and API Gateways to be redeployed when upgrading eoapi-cdk. Being that both services are stateless, this should not cause issues unless a developer has made manual modifications or manual references to the deployed lambdas/api gateways.